Rearrange lasagna.html and add inline-blocks
This will change the layout of lasagna.html so that the subtitle is
just below the title. It will also change the lasagna image so that it
is an inline-block. This will also merge the nutritional info and prep
times into one block which was also changed to an inline block.
This will also remove the top margin on this block.